api-userstate: dto и model перенесены в модуль core
