+++ .gitignore
@@ -7,6 +7,7 @@ target/
 .mvn
 mvnw
 mvnw.cmd
+.jpb/
 
 ### STS ###
 .apt_generated
