+++ core/src/main/java/ru/teamtwo/core/dtos/user/CartItemArrayDto.java
@@ -1,10 +1,9 @@
-package ru.teamtwo.website.dtos.user;
+package ru.teamtwo.core.dtos.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
 import java.util.Set;
 
 @Data
