+++ core/src/main/java/ru/teamtwo/core/dtos/user/CustomerDto.java
@@ -1,9 +1,9 @@
-package ru.teamtwo.website.dtos.user;
+package ru.teamtwo.core.dtos.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.teamtwo.website.model.user.Customer;
+import ru.teamtwo.core.models.user.Customer;
 
 import java.io.Serializable;
