+++ core/src/main/java/ru/teamtwo/core/dtos/user/OrderDto.java
@@ -1,8 +1,9 @@
-package ru.teamtwo.telegrambot.dtos;
+package ru.teamtwo.core.dtos.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import ru.teamtwo.core.models.user.Order;
 
 import java.io.Serializable;
 import java.time.LocalDate;
@@ -14,4 +15,10 @@ public class OrderDto implements Serializable {
     private Integer id;
     private Integer customerId;
     private LocalDate createdOn;
+
+    public OrderDto(Order entity) {
+        this.id = entity.getId();
+        this.customerId = entity.getCustomer().getId();
+        this.createdOn = entity.getCreatedOn();
+    }
 }
