+++ core/src/main/java/ru/teamtwo/core/dtos/user/OrderItemDto.java
@@ -1,8 +1,9 @@
-package ru.teamtwo.telegrambot.dtos;
+package ru.teamtwo.core.dtos.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import ru.teamtwo.core.models.user.OrderItem;
 
 import java.io.Serializable;
 
@@ -14,4 +15,11 @@ public class OrderItemDto implements Serializable {
     private Integer orderId;
     private Integer productOfferId;
     private Integer quantity;
+
+    public OrderItemDto(OrderItem entity) {
+        this.id = entity.getId();
+        this.orderId = entity.getOrder().getId();
+        this.productOfferId = entity.getProductOffer().getId();
+        this.quantity = entity.getQuantity();
+    }
 }
