+++ core/src/main/java/ru/teamtwo/core/models/Product.java
@@ -1,4 +1,4 @@
-package ru.teamtwo.website.model;
+package ru.teamtwo.core.models;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -13,8 +13,8 @@ import javax.persistence.Table;
 @Getter
 @Setter
 @Entity
-@Table(name = "store", schema = "marketplace")
-public class Store {
+@Table(name = "product", schema = "marketplace")
+public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
@@ -22,6 +22,21 @@ public class Store {
 
     @Column(name = "name", nullable = false)
     private String name;
+
+    @Column(name = "category", nullable = false)
+    private String category;
+
+    @Column(name = "model", nullable = false)
+    private String model;
+
+    @Column(name = "manufacturer", nullable = false)
+    private String manufacturer;
+
+    @Column(name = "description", nullable = false)
+    private String description;
+
+    @Column(name = "price", nullable = false)
+    private Integer price;
 
     @Column(name = "rating", nullable = false)
     private Integer rating;
