+++ core/src/main/java/ru/teamtwo/core/models/Store.java
@@ -1,6 +1,5 @@
-package ru.teamtwo.website.model;
+package ru.teamtwo.core.models;
 
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,14 +8,13 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.Lob;
 import javax.persistence.Table;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "product", schema = "marketplace")
-public class Product {
+@Table(name = "store", schema = "marketplace")
+public class Store {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
@@ -24,21 +22,6 @@ public class Product {
 
     @Column(name = "name", nullable = false)
     private String name;
-
-    @Column(name = "category", nullable = false)
-    private String category;
-
-    @Column(name = "model", nullable = false)
-    private String model;
-
-    @Column(name = "manufacturer", nullable = false)
-    private String manufacturer;
-
-    @Column(name = "description", nullable = false)
-    private String description;
-
-    @Column(name = "price", nullable = false)
-    private Integer price;
 
     @Column(name = "rating", nullable = false)
     private Integer rating;
