+++ core/src/main/java/ru/teamtwo/core/models/user/Order.java
@@ -1,12 +1,10 @@
-package ru.teamtwo.website.model.user;
+package ru.teamtwo.core.models.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import ru.teamtwo.website.dtos.user.CartItemDto;
-import ru.teamtwo.website.dtos.user.OrderDto;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
