+++ telegrambot/pom.xml
@@ -42,6 +42,11 @@
             <artifactId>spring-boot-starter-webflux</artifactId>
             <version>2.6.5</version>
         </dependency>
+        <dependency>
+            <groupId>ru.teamtwo</groupId>
+            <artifactId>core</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
     </dependencies>
 
     <build>
