--- telegrambot/src/main/java/ru/teamtwo/telegrambot/dtos/CustomerDto.java
@@ -1,16 +1,0 @@
-package ru.teamtwo.telegrambot.dtos;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class CustomerDto implements Serializable {
-    private Integer id;
-    private String name;
-    private String address;
-}
