+++ telegrambot/src/test/java/ru/teamtwo/telegrambot/RESTTests.java
@@ -6,7 +6,7 @@ import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.telegram.telegrambots.meta.api.objects.User;
-import ru.teamtwo.telegrambot.dtos.CustomerDto;
+import ru.teamtwo.core.dtos.user.CustomerDto;
 import ru.teamtwo.telegrambot.model.UserState;
 import ru.teamtwo.telegrambot.service.TelegramBotRESTHandler;
