+++ website/src/main/java/ru/teamtwo/website/WebsiteApplication.java
@@ -2,8 +2,12 @@ package ru.teamtwo.website;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackages = { "ru.teamtwo.*" })
+@EntityScan("ru.teamtwo.*")
 public class WebsiteApplication {
 
     public static void main(String[] args) {
