+++ website/src/main/java/ru/teamtwo/website/controller/MarketplaceController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import ru.teamtwo.website.dtos.ProductDTO;
+import ru.teamtwo.core.dtos.ProductDTO;
 import ru.teamtwo.website.repository.ProductOfferRepository;
 
 import java.util.List;
