+++ website/src/main/java/ru/teamtwo/website/controller/ProductOfferController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ru.teamtwo.website.dtos.ProductDTO;
+import ru.teamtwo.core.dtos.ProductDTO;
 import ru.teamtwo.website.repository.ProductOfferRepository;
 
 @Slf4j
