+++ website/src/main/java/ru/teamtwo/website/controller/user/CartItemController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
-import ru.teamtwo.website.dtos.user.CartItemArrayDto;
-import ru.teamtwo.website.dtos.user.CartItemDto;
-import ru.teamtwo.website.model.user.CartItem;
+import ru.teamtwo.core.dtos.user.CartItemArrayDto;
+import ru.teamtwo.core.dtos.user.CartItemDto;
+import ru.teamtwo.core.models.user.CartItem;
 import ru.teamtwo.website.repository.ProductRepository;
 import ru.teamtwo.website.repository.user.CartItemRepository;
 import ru.teamtwo.website.repository.user.CustomerRepository;
