+++ website/src/main/java/ru/teamtwo/website/controller/user/OrderItemController.java
@@ -11,10 +11,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
-import ru.teamtwo.website.dtos.user.OrderItemDto;
-import ru.teamtwo.website.model.ProductOffer;
-import ru.teamtwo.website.model.user.Order;
-import ru.teamtwo.website.model.user.OrderItem;
+import ru.teamtwo.core.dtos.user.OrderItemDto;
+import ru.teamtwo.core.models.ProductOffer;
+import ru.teamtwo.core.models.user.Order;
+import ru.teamtwo.core.models.user.OrderItem;
 import ru.teamtwo.website.repository.ProductOfferRepository;
 import ru.teamtwo.website.repository.user.OrderItemRepository;
 import ru.teamtwo.website.repository.user.OrderRepository;
