+++ website/src/main/java/ru/teamtwo/website/repository/ProductOfferRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
-import ru.teamtwo.website.model.ProductOffer;
+import ru.teamtwo.core.models.ProductOffer;
 
 public interface ProductOfferRepository extends PagingAndSortingRepository<ProductOffer, Long> {
     @Query(value = "SELECT po FROM ProductOffer po where po.product.name LIKE %:productName%")
