+++ website/src/main/java/ru/teamtwo/website/repository/ProductRepository.java
@@ -1,7 +1,7 @@
 package ru.teamtwo.website.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.teamtwo.website.model.Product;
+import ru.teamtwo.core.models.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Integer> {
 }
