+++ website/src/main/java/ru/teamtwo/website/repository/StoreRepository.java
@@ -1,7 +1,7 @@
 package ru.teamtwo.website.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.teamtwo.website.model.Store;
+import ru.teamtwo.core.models.Store;
 
 public interface StoreRepository extends JpaRepository<Store, Long> {
 }
