+++ website/src/main/java/ru/teamtwo/website/repository/user/CartItemRepository.java
@@ -1,7 +1,7 @@
 package ru.teamtwo.website.repository.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.teamtwo.website.model.user.CartItem;
+import ru.teamtwo.core.models.user.CartItem;
 
 import java.util.Set;
