+++ website/src/main/java/ru/teamtwo/website/repository/user/CustomerRepository.java
@@ -1,7 +1,7 @@
 package ru.teamtwo.website.repository.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.teamtwo.website.model.user.Customer;
+import ru.teamtwo.core.models.user.Customer;
 
 public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 }
