+++ website/src/main/java/ru/teamtwo/website/repository/user/OrderItemRepository.java
@@ -1,7 +1,7 @@
 package ru.teamtwo.website.repository.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.teamtwo.website.model.user.OrderItem;
+import ru.teamtwo.core.models.user.OrderItem;
 
 public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
 }
