+++ website/src/main/java/ru/teamtwo/website/repository/user/OrderRepository.java
@@ -1,7 +1,7 @@
 package ru.teamtwo.website.repository.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.teamtwo.website.model.user.Order;
+import ru.teamtwo.core.models.user.Order;
 
 public interface OrderRepository extends JpaRepository<Order, Integer> {
 }
